feat(typing): Move `ChartType`, `is_chart_type` to `alt.typing`
These aren't introduced until `5.4.0`, so this is keeps the top-level from growing +2

https://github.com/vega/altair/pull/3515#discussion_r1703165544
