+++ altair/__init__.py
@@ -50,7 +50,6 @@ __all__ = [
     "ChainedWhen",
     "Chart",
     "ChartDataType",
-    "ChartType",
     "Color",
     "ColorDatum",
     "ColorDef",
@@ -611,7 +610,6 @@ __all__ = [
     "expr",
     "graticule",
     "hconcat",
-    "is_chart_type",
     "jupyter",
     "layer",
     "limit_rows",
