+++ altair/utils/_transformed_data.py
@@ -31,8 +31,8 @@ from altair.utils._vegafusion_data import get_inline_tables, import_vegafusion
 from altair.utils.schemapi import Undefined
 
 if TYPE_CHECKING:
+    from altair.typing import ChartType
     from altair.utils.core import DataFrameLike
-    from altair.vegalite.v5.api import ChartType
 
 Scope: TypeAlias = Tuple[int, ...]
 FacetMapping: TypeAlias = Dict[Tuple[str, Scope], Tuple[str, Scope]]
