+++ altair/vegalite/v5/api.py
@@ -145,7 +145,6 @@ __all__ = [
     "ChainedWhen",
     "Chart",
     "ChartDataType",
-    "ChartType",
     "ConcatChart",
     "DataType",
     "FacetChart",
@@ -174,7 +173,6 @@ __all__ = [
     "condition",
     "graticule",
     "hconcat",
-    "is_chart_type",
     "layer",
     "mixins",
     "param",
